sentence_transformer_streamlit_build/requirements.txt
streamlit
pandas
sentence-transformers
torch
transformers
scikit-learn
numpy
matplotlib
seaborn
plotly
python-docx
pdf2image
pytesseract